+++ hardware/openzwave/value_classes/ValueSchedule.h
@@ -40,11 +40,12 @@ namespace OpenZWave
 	class Node;
 
 	/** \brief Schedule sent to/received from a node.
+	 * \ingroup ValueID
 	 */
 	class ValueSchedule: public Value
 	{
 	public:
-		ValueSchedule( uint32 const _homeId, uint8 const _nodeId, ValueID::ValueGenre const _genre, uint8 const _commandClassId, uint8 const _instance, uint8 const _index, string const& _label, string const& _units, bool const _readOnly, bool const _writeOnly, uint8 const _pollIntensity );
+		ValueSchedule( uint32 const _homeId, uint8 const _nodeId, ValueID::ValueGenre const _genre, uint8 const _commandClassId, uint8 const _instance, uint16 const _index, string const& _label, string const& _units, bool const _readOnly, bool const _writeOnly, uint8 const _pollIntensity );
 		ValueSchedule();
 		virtual ~ValueSchedule(){}
